Extract test setup function to own file
